HSEARCH-4563 Update tests for query predicates
