+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/search/predicate/QueryStringPredicateSpecificsIT.java
@@ -9,318 +9,316 @@ package org.hibernate.search.integrationtest.backend.tck.search.predicate;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.hibernate.search.util.impl.integrationtest.common.assertion.SearchResultAssert.assertThatQuery;
 
-import java.util.function.Function;
+import java.util.List;
+import java.util.function.Supplier;
 
 import org.hibernate.search.engine.backend.common.DocumentReference;
-import org.hibernate.search.engine.backend.document.IndexObjectFieldReference;
-import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaElement;
-import org.hibernate.search.engine.backend.document.model.dsl.IndexSchemaObjectField;
-import org.hibernate.search.engine.backend.types.ObjectStructure;
-import org.hibernate.search.engine.search.common.BooleanOperator;
+import org.hibernate.search.engine.search.common.RewriteMethod;
+import org.hibernate.search.engine.search.predicate.SearchPredicate;
+import org.hibernate.search.engine.search.predicate.dsl.QueryStringPredicateFieldStep;
+import org.hibernate.search.engine.search.predicate.dsl.QueryStringPredicateOptionsStep;
+import org.hibernate.search.engine.search.predicate.dsl.SearchPredicateFactory;
 import org.hibernate.search.engine.search.query.SearchQuery;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.configuration.DefaultAnalysisDefinitions;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.types.AnalyzedStringFieldTypeDescriptor;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.types.KeywordStringFieldTypeDescriptor;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.util.SimpleFieldModel;
-import org.hibernate.search.integrationtest.backend.tck.testsupport.util.extension.SearchSetupHelper;
 import org.hibernate.search.util.common.SearchException;
-import org.hibernate.search.util.impl.integrationtest.mapper.stub.BulkIndexer;
-import org.hibernate.search.util.impl.integrationtest.mapper.stub.SimpleMappedIndex;
 import org.hibernate.search.util.impl.integrationtest.mapper.stub.StubMappingScope;
-import org.hibernate.search.util.impl.test.annotation.PortedFromSearch5;
-import org.hibernate.search.util.impl.test.annotation.TestForIssue;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 
-class QueryStringPredicateSpecificsIT {
+class QueryStringPredicateSpecificsIT extends AbstractBaseQueryStringPredicateSpecificsIT<QueryStringPredicateFieldStep<?>> {
 
-	private static final String DOCUMENT_1 = "document1";
-	private static final String DOCUMENT_2 = "document2";
-	private static final String DOCUMENT_3 = "document3";
-	private static final String DOCUMENT_4 = "document4";
-	private static final String DOCUMENT_5 = "document5";
-	private static final String EMPTY = "empty";
-
-	private static final String TERM_1 = "word";
-	private static final String TERM_2 = "panda";
-	private static final String TERM_3 = "room";
-	private static final String PHRASE_WITH_TERM_2 = "panda breeding";
-	private static final String PHRASE_WITH_TERM_4 = "elephant john";
-	private static final String PREFIX_FOR_TERM_1_AND_TERM_6 = "wor";
-	private static final String PREFIX_FOR_TERM_6 = "worl";
-	private static final String PREFIX_FOR_TERM_1_AND_TERM_6_DIFFERENT_CASE = "Wor";
-	private static final String PREFIX_FOR_TERM_6_DIFFERENT_CASE = "Worl";
-	private static final String TEXT_TERM_1_AND_TERM_2 = "Here I was, feeding my panda, and the crowd had no word.";
-	private static final String TEXT_TERM_1_AND_TERM_3 = "Without a word, he went out of the room.";
-	private static final String TEXT_TERM_2_IN_PHRASE = "I admired her for her panda breeding expertise.";
-	private static final String TEXT_TERM_4_IN_PHRASE_SLOP_2 = "An elephant ran past John.";
-	private static final String TEXT_TERM_1_EDIT_DISTANCE_1_OR_TERM_6 = "I came to the world in a dumpster.";
-
-	@RegisterExtension
-	public static final SearchSetupHelper setupHelper = SearchSetupHelper.create();
-
-	private static final SimpleMappedIndex<IndexBinding> index = SimpleMappedIndex.of( IndexBinding::new );
-
-	private static final DataSet dataSet = new DataSet();
-
-	@BeforeAll
-	static void setup() {
-		setupHelper.start().withIndexes( index ).setup();
-
-		BulkIndexer indexer = index.bulkIndexer();
-		dataSet.contribute( indexer );
-		indexer.join();
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-2678")
-	@PortedFromSearch5(original = "org.hibernate.search.test.dsl.queryStringDSLTest.testqueryString")
-	void booleanOperators() {
+	@ParameterizedTest
+	@MethodSource("rewriteMethodOptions")
+	void rewriteMethodOptions(RewriteMethod rewriteMethod, Integer n) {
 		StubMappingScope scope = index.createScope();
 		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-		Function<String, SearchQuery<DocumentReference>> createQuery = queryString -> scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath ).matching( queryString ) )
-				.toQuery();
 
-		assertThatQuery( createQuery.apply( TERM_1 + " " + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_3 );
+		QueryStringPredicateOptionsStep<?> optionsStep = scope.predicate().queryString()
+				.field( absoluteFieldPath )
+				.matching( TERM_1 );
 
-		assertThatQuery( createQuery.apply( TERM_1 + " | " + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_3 );
+		SearchPredicate predicate =
+				( n == null ? optionsStep.rewriteMethod( rewriteMethod ) : optionsStep.rewriteMethod( rewriteMethod, n ) )
+						.toPredicate();
 
-		assertThatQuery( createQuery.apply( "+" + TERM_1 + " +" + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1 );
-
-		assertThatQuery( createQuery.apply( "-" + TERM_1 + " +" + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_3 );
-
-		assertThatQuery( createQuery.apply( TERM_1 + " + -" + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_2 );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-3844") // Used to throw NPE
-	void nonAnalyzedField() {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().nonAnalyzedField.relativeFieldName;
-		Function<String, SearchQuery<DocumentReference>> createQuery = queryString -> scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath ).matching( queryString ) )
-				.toQuery();
-
-		assertThatQuery( createQuery.apply( TERM_1 + " " + TERM_2 ) )
-				.hasNoHits();
-
-		assertThatQuery( createQuery.apply( TERM_1 + " | " + TERM_2 ) )
-				.hasNoHits();
-
-		assertThatQuery( createQuery.apply( TERM_1 + " +" + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_3 );
-
-		assertThatQuery( createQuery.apply( "-" + TERM_1 + " +" + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_3 );
-
-		assertThatQuery( createQuery.apply( TERM_1 + " + -" + TERM_2 ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_2 );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-2678")
-	@PortedFromSearch5(original = "org.hibernate.search.test.dsl.queryStringDSLTest.testqueryString")
-	void defaultOperator() {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-		SearchQuery<DocumentReference> query;
-
-		query = scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath )
-						.matching( TERM_1 + " " + TERM_2 ) )
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( predicate )
 				.toQuery();
 		assertThatQuery( query )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_3 );
-
-		query = scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath )
-						.matching( TERM_1 + " " + TERM_2 )
-						.defaultOperator( BooleanOperator.OR ) )
-				.toQuery();
-		assertThatQuery( query )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_3 );
-
-		query = scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath )
-						.matching( TERM_1 + " " + TERM_2 )
-						.defaultOperator( BooleanOperator.AND ) )
-				.toQuery();
-		assertThatQuery( query )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1 );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-2678")
-	@PortedFromSearch5(original = "org.hibernate.search.test.dsl.queryStringDSLTest.testqueryString")
-	void phrase() {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-		Function<String, SearchQuery<DocumentReference>> createQuery = queryString -> scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath ).matching( queryString ) )
-				.toQuery();
-
-		assertThatQuery( createQuery.apply( "\"" + PHRASE_WITH_TERM_2 + "\"" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_3 );
-
-		assertThatQuery( createQuery.apply( TERM_3 + " \"" + PHRASE_WITH_TERM_2 + "\"" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_2, DOCUMENT_3 );
-
-		// Slop
-		assertThatQuery( createQuery.apply( "\"" + PHRASE_WITH_TERM_4 + "\"" ) )
-				.hasNoHits();
-		assertThatQuery( createQuery.apply( "\"" + PHRASE_WITH_TERM_4 + "\"~1" ) )
-				.hasNoHits();
-		assertThatQuery( createQuery.apply( "\"" + PHRASE_WITH_TERM_4 + "\"~2" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_4 );
-		assertThatQuery( createQuery.apply( "\"" + PHRASE_WITH_TERM_4 + "\"~3" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_4 );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-2678")
-	@PortedFromSearch5(original = "org.hibernate.search.test.dsl.queryStringDSLTest.testFuzzy")
-	void fuzzy() {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-		Function<String, SearchQuery<DocumentReference>> createQuery = queryString -> scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath ).matching( queryString ) )
-				.toQuery();
-
-		assertThatQuery( createQuery.apply( TERM_1 ) )
 				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2 );
-
-		assertThatQuery( createQuery.apply( TERM_1 + "~1" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_5 );
-
-		assertThatQuery( createQuery.apply( TERM_1 + "~2" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_5 );
 	}
 
-	@Test
-	void prefix() {
+	@ParameterizedTest
+	@MethodSource("rewriteMethodOptions")
+	void rewriteOptions_exceptions(RewriteMethod rewrite, Integer n) {
 		StubMappingScope scope = index.createScope();
 		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
 
-		Function<String, SearchQuery<DocumentReference>> createQuery = queryString -> scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath ).matching( queryString ) )
-				.toQuery();
+		QueryStringPredicateOptionsStep<?> optionsStep = scope.predicate().queryString()
+				.field( absoluteFieldPath )
+				.matching( TERM_1 );
 
-		assertThatQuery( createQuery.apply( PREFIX_FOR_TERM_1_AND_TERM_6 + "*" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_5 );
-
-		assertThatQuery( createQuery.apply( PREFIX_FOR_TERM_6 + "*" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_5 );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = { "HSEARCH-3612", "HSEARCH-3845" })
-	void prefix_normalizePrefixTerm() {
-		StubMappingScope scope = index.createScope();
-		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
-
-		Function<String, SearchQuery<DocumentReference>> createQuery = queryString -> scope.query()
-				.where( f -> f.queryString().field( absoluteFieldPath ).matching( queryString ) )
-				.toQuery();
-
-		assertThatQuery( createQuery.apply( PREFIX_FOR_TERM_1_AND_TERM_6_DIFFERENT_CASE + "*" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_1, DOCUMENT_2, DOCUMENT_5 );
-
-		assertThatQuery( createQuery.apply( PREFIX_FOR_TERM_6_DIFFERENT_CASE + "*" ) )
-				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_5 );
-	}
-
-	@Test
-	void incompatibleNestedPaths() {
-		String fieldInRootPath = index.binding().analyzedStringField1.relativeFieldName;
-		String fieldInNestedPath = index.binding().nested.fieldPath();
-		assertThatThrownBy( () -> index.createScope()
-				.predicate().queryString().field( fieldInNestedPath ).field( fieldInRootPath ) )
+		assertThatThrownBy(
+				() -> ( n != null ? optionsStep.rewriteMethod( rewrite ) : optionsStep.rewriteMethod( rewrite, 10 ) )
+						.toPredicate() )
 				.isInstanceOf( SearchException.class )
-				.hasMessageContainingAll(
-						"Invalid target fields:",
-						"fields [" + fieldInNestedPath + ", " + fieldInRootPath +
-								"] are in different nested documents (field 'nested' vs. index schema root)",
-						"All target fields must be in the same document"
-				);
+				.hasMessageContaining(
+						n == null
+								? ( "Cannot use rewrite method '"
+										+ rewrite
+										+ "': this method does not accept parameter 'n', but it was specified."
+										+ " Use another version of the rewrite(...) method that does not accept parameter 'n'." )
+								: ( "Cannot use rewrite method '"
+										+ rewrite
+										+ "': this method requires parameter 'n', which was not specified."
+										+ " Use another version of the rewrite(...) method that accepts parameter 'n'." ) );
 	}
 
-	private static class IndexBinding {
-		final SimpleFieldModel<String> analyzedStringField1;
-		final SimpleFieldModel<String> nonAnalyzedField;
-
-		final ObjectFieldBinding nested;
-
-		IndexBinding(IndexSchemaElement root) {
-			analyzedStringField1 = SimpleFieldModel.mapperWithOverride( AnalyzedStringFieldTypeDescriptor.INSTANCE,
-					c -> c.asString().analyzer( DefaultAnalysisDefinitions.ANALYZER_STANDARD_ENGLISH.name )
-			)
-					.map( root, "analyzedString1" );
-			// A field without any analyzer or normalizer
-			nonAnalyzedField = SimpleFieldModel.mapper( KeywordStringFieldTypeDescriptor.INSTANCE )
-					.map( root, "nonAnalyzed" );
-			nested = ObjectFieldBinding.create( root, null, "nested", ObjectStructure.NESTED );
-		}
+	public static List<? extends Arguments> rewriteMethodOptions() {
+		return List.of(
+				Arguments.of( RewriteMethod.CONSTANT_SCORE, null ),
+				Arguments.of( RewriteMethod.CONSTANT_SCORE_BOOLEAN, null ),
+				Arguments.of( RewriteMethod.SCORING_BOOLEAN, null ),
+				Arguments.of( RewriteMethod.TOP_TERMS_BLENDED_FREQS_N, 2 ),
+				Arguments.of( RewriteMethod.TOP_TERMS_BOOST_N, 2 ),
+				Arguments.of( RewriteMethod.TOP_TERMS_N, 2 )
+		);
 	}
 
-	static class ObjectFieldBinding {
-		final IndexObjectFieldReference reference;
-		final String absolutePath;
+	@ParameterizedTest
+	@MethodSource("phraseSlop")
+	void phraseSlop(String phrase, int slop, String docId) {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
 
-		final SimpleFieldModel<String> field;
-
-		static ObjectFieldBinding create(IndexSchemaElement parent, String parentAbsolutePath, String relativeFieldName,
-				ObjectStructure structure) {
-			IndexSchemaObjectField objectField = parent.objectField( relativeFieldName, structure );
-			String absolutePath = parentAbsolutePath == null ? relativeFieldName : parentAbsolutePath + "." + relativeFieldName;
-			return new ObjectFieldBinding( objectField, absolutePath );
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( phrase )
+						.phraseSlop( slop ) )
+				.toQuery();
+		if ( docId == null ) {
+			assertThatQuery( query )
+					.hasNoHits();
 		}
-
-		ObjectFieldBinding(IndexSchemaObjectField objectField, String absolutePath) {
-			reference = objectField.toReference();
-			this.absolutePath = absolutePath;
-			field = SimpleFieldModel.mapper( AnalyzedStringFieldTypeDescriptor.INSTANCE )
-					.map( objectField, "field" );
-		}
-
-		String fieldPath() {
-			return absolutePath + "." + field.relativeFieldName;
+		else {
+			assertThatQuery( query )
+					.hasDocRefHitsAnyOrder( index.typeName(), docId );
 		}
 	}
 
-	private static final class DataSet extends AbstractPredicateDataSet {
-		public DataSet() {
-			super( null );
+	public static List<? extends Arguments> phraseSlop() {
+		// Here I was, feeding my panda, and the crowd had no word
+		String phraseReorderedWords = "\"had crowd\"";
+		String phraseWithoutTwoWordsInTheMiddle = "\"crowd word\"";
+
+		return List.of(
+				// default value means phrases should be exact as in text:
+				Arguments.of( phraseReorderedWords, 0, null ),
+				// Transposed terms have a slop of 2
+				Arguments.of( phraseReorderedWords, 2, DOCUMENT_1 ),
+				// not enough so should not match anything
+				Arguments.of( phraseReorderedWords, 1, null ),
+				Arguments.of( phraseWithoutTwoWordsInTheMiddle, 1, null ),
+				Arguments.of( phraseWithoutTwoWordsInTheMiddle, 2, DOCUMENT_1 ),
+				Arguments.of( phraseWithoutTwoWordsInTheMiddle, 5, DOCUMENT_1 )
+		);
+	}
+
+	@ParameterizedTest
+	@MethodSource("allowLeadingWildcard")
+	void allowLeadingWildcard(String phrase, Boolean allow, String docId) {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		QueryStringPredicateOptionsStep<?> step = scope.predicate().queryString()
+				.field( absoluteFieldPath )
+				.matching( phrase );
+
+		Supplier<SearchPredicate> supplier =
+				() -> allow != null ? step.allowLeadingWildcard( allow ).toPredicate() : step.toPredicate();
+
+		if ( docId == null ) {
+			assertThatThrownBy( () -> scope.query()
+					.where( supplier.get() ).toQuery().fetchAll() )
+					// Lucene will throw org.apache.lucene.queryparser.classic.ParseException
+					// Elasticsearch will fail with SearchException reported by the backend: "reason": "Failed to parse query ...
+					.isInstanceOf( RuntimeException.class );
+			return;
 		}
 
-		public void contribute(BulkIndexer indexer) {
-			indexer.add( DOCUMENT_1, document -> {
-				document.addValue( index.binding().nonAnalyzedField.reference, TEXT_TERM_1_AND_TERM_2 );
-				document.addValue( index.binding().analyzedStringField1.reference, TEXT_TERM_1_AND_TERM_2 );
-			} )
-					.add( DOCUMENT_2, document -> {
-						document.addValue( index.binding().nonAnalyzedField.reference, TERM_1 );
-						document.addValue( index.binding().analyzedStringField1.reference, TEXT_TERM_1_AND_TERM_3 );
-					} )
-					.add( DOCUMENT_3, document -> {
-						document.addValue( index.binding().nonAnalyzedField.reference, TERM_2 );
-						document.addValue( index.binding().analyzedStringField1.reference, TEXT_TERM_2_IN_PHRASE );
-					} )
-					.add( DOCUMENT_4, document -> {
-						document.addValue( index.binding().analyzedStringField1.reference, TEXT_TERM_4_IN_PHRASE_SLOP_2 );
-					} )
-					.add( DOCUMENT_5, document -> {
-						document.addValue( index.binding().analyzedStringField1.reference,
-								TEXT_TERM_1_EDIT_DISTANCE_1_OR_TERM_6 );
-					} )
-					.add( EMPTY, document -> {} );
+		SearchPredicate predicate = supplier.get();
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( predicate )
+				.toQuery();
+		assertThatQuery( query )
+				.hasDocRefHitsAnyOrder( index.typeName(), docId );
+	}
+
+	public static List<? extends Arguments> allowLeadingWildcard() {
+		// "An elephant ran past John."
+		return List.of(
+				Arguments.of( "?lephant", null, DOCUMENT_4 ),
+				Arguments.of( "?lephant", true, DOCUMENT_4 ),
+				Arguments.of( "?lephant", false, null ),
+
+				Arguments.of( "*hant", null, DOCUMENT_4 ),
+				Arguments.of( "*hant", true, DOCUMENT_4 ),
+				Arguments.of( "*hant", false, null ),
+
+				Arguments.of( "ele?hant", null, DOCUMENT_4 ),
+				Arguments.of( "ele?hant", true, DOCUMENT_4 ),
+				Arguments.of( "ele?hant", false, DOCUMENT_4 ),
+
+				Arguments.of( "el*nt", null, DOCUMENT_4 ),
+				Arguments.of( "el*nt", true, DOCUMENT_4 ),
+				Arguments.of( "el*nt", false, DOCUMENT_4 )
+		);
+	}
+
+	@ParameterizedTest
+	@MethodSource("minimumShouldMatch")
+	void minimumShouldMatch(String phrase, int min, String docId, String[] otherDocIds) {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( phrase )
+						.minimumShouldMatchNumber( min ) )
+				.toQuery();
+		if ( docId == null ) {
+			assertThatQuery( query )
+					.hasNoHits();
+		}
+		else {
+			assertThatQuery( query )
+					.hasDocRefHitsAnyOrder( index.typeName(), docId, otherDocIds );
 		}
 	}
+
+	@ParameterizedTest
+	@MethodSource("minimumShouldMatch")
+	void minimumShouldMatch_dsl(String phrase, int min, String docId, String[] otherDocIds) {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( phrase )
+						.minimumShouldMatch().ifMoreThan( 0 )
+						.thenRequireNumber( min )
+						.end() )
+				.toQuery();
+		if ( docId == null ) {
+			assertThatQuery( query )
+					.hasNoHits();
+		}
+		else {
+			assertThatQuery( query )
+					.hasDocRefHitsAnyOrder( index.typeName(), docId, otherDocIds );
+		}
+	}
+
+	public static List<? extends Arguments> minimumShouldMatch() {
+		// "Here I was, feeding my panda, and the crowd had no word."
+		return List.of(
+				Arguments.of( "feeding my panda", 1, DOCUMENT_1, new String[] { DOCUMENT_3 } ),
+				Arguments.of( "feeding my panda", 2, DOCUMENT_1, new String[] { } ),
+				Arguments.of( "feeding my panda", 3, DOCUMENT_1, new String[] { } ),
+				Arguments.of( "feeding my panda crowd had", 5, DOCUMENT_1, new String[] { } ),
+				Arguments.of( "some other panda less matches", 5, null, new String[] { } )
+		);
+	}
+
+	@ParameterizedTest
+	@MethodSource("minimumShouldMatchPercent")
+	void minimumShouldMatch_percent(String phrase, int min, String docId, String[] otherDocIds) {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( phrase )
+						.minimumShouldMatchPercent( min ) )
+				.toQuery();
+		if ( docId == null ) {
+			assertThatQuery( query )
+					.hasNoHits();
+		}
+		else {
+			assertThatQuery( query )
+					.hasDocRefHitsAnyOrder( index.typeName(), docId, otherDocIds );
+		}
+	}
+
+	public static List<? extends Arguments> minimumShouldMatchPercent() {
+		// "Here I was, feeding my panda, and the crowd had no word."
+		return List.of(
+				Arguments.of( "feeding my panda", 34, DOCUMENT_1, new String[] { DOCUMENT_3 } ),
+				Arguments.of( "feeding my panda", 67, DOCUMENT_1, new String[] { } ),
+				Arguments.of( "feeding my panda", 100, DOCUMENT_1, new String[] { } ),
+				Arguments.of( "feeding my panda crowd had", 100, DOCUMENT_1, new String[] { } ),
+				Arguments.of( "some other panda less matches", 100, null, new String[] { } )
+		);
+	}
+
+	@Test
+	void enablePositionIncrements_disabled() {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( "\"out of the room\"" )
+						.enablePositionIncrements( false ) )
+				.toQuery();
+
+		assertThatQuery( query )
+				.hasNoHits();
+	}
+
+	@Test
+	void enablePositionIncrements_enabled() {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( "\"out of the room\"" )
+						.enablePositionIncrements( true ) )
+				.toQuery();
+
+		assertThatQuery( query )
+				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_2 );
+	}
+
+	@Test
+	void enablePositionIncrements_default() {
+		StubMappingScope scope = index.createScope();
+		String absoluteFieldPath = index.binding().analyzedStringField1.relativeFieldName;
+
+		SearchQuery<DocumentReference> query = scope.query()
+				.where( f -> f.queryString()
+						.field( absoluteFieldPath )
+						.matching( "\"out of the room\"" ) )
+				.toQuery();
+
+		assertThatQuery( query )
+				.hasDocRefHitsAnyOrder( index.typeName(), DOCUMENT_2 );
+	}
+
+	@Override
+	QueryStringPredicateFieldStep<?> predicate(SearchPredicateFactory f) {
+		return f.queryString();
+	}
+
 }
